requirements.txt
google-generativeai
pyautogen